Moving inline implementations in concrete/ to below the class interfaces
git-svn-id: https://cppannotations.svn.sourceforge.net/svnroot/cppannotations/trunk@33 f6dd340e-d3f9-0310-b409-bdd246841980

+++ yo/concrete/examples/pipe.cc
@@ -43,7 +43,7 @@
     {
         writeOnly();
 
-        for (int idx = 0; idx < n; idx++)
+        for (size_t idx = 0; idx < n; idx++)
             redirect(d_fd[WRITE], fd[idx]);
 
         close(d_fd[WRITE]);
